% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsurf-class.R
\name{coef.effectsurf}
\alias{coef.effectsurf}
\title{Extract coefficients (estimates) from an effectsurf object}
\usage{
\method{coef}{effectsurf}(object, ...)
}
\arguments{
\item{object}{An \code{effectsurf} object.}

\item{...}{Ignored.}
}
\value{
Numeric vector of estimates.
}
\description{
Extract coefficients (estimates) from an effectsurf object
}
