% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{expit}
\alias{expit}
\title{Expit (inverse logit) function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector on (0, 1) scale.
}
\description{
Computes the inverse of the logit transformation:
\eqn{expit(x) = \frac{e^x}{1 + e^x}}{expit(x) = exp(x) / (1 + exp(x))}.
}
\examples{
expit(0)     # 0.5
expit(-Inf)  # 0
expit(Inf)   # 1
}
