% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsurf-class.R
\name{is_effectsurf}
\alias{is_effectsurf}
\title{Check if an object is an effectsurf object}
\usage{
is_effectsurf(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical.
}
\description{
Check if an object is an effectsurf object
}
